src/app/requirements.txt
flask
grpcio
requests
protobuf
